fix: 메인페이지 분리후 suspense로 감싸 줌
