+++ components/index.ts
@@ -33,3 +33,4 @@ export { default as GoogleAnalytics } from './google-analytics'
 export { default as Filter } from './filter'
 export { default as DialogShare } from './dialog-share'
 export { default as Tooltip } from './tooltip'
+export { default as MainPageContent } from './main-page-content'
